Increased lpeg vm instruction size by 2 bytes in order to accomodate a larger aux field (was byte, now short).  That means we can have a longer lookbehind length limit.  Current limit, MAXBEHIND, is set to 8000.  Do we have any rationale for setting higher?  Or lower?

+++ src/lpvm.h
@@ -41,8 +41,8 @@ typedef enum Opcode {
 
 typedef union Instruction {
   struct Inst {
-    byte code;
-    byte aux;
+    short code;			/* Rosie (was byte) */
+    short aux;			/* Rosie (was byte) */
     short key;
   } i;
   int offset;
